feat: add genre entity
