+++ data/src/main/java/com/madrid/data/dataSource/local/mappers/GenreMappers.kt
@@ -1,22 +1,14 @@
 package com.madrid.data.dataSource.local.mappers
 
 import com.madrid.data.dataSource.local.table.MovieGenreTable
-import com.madrid.data.dataSource.local.table.SeriesGenreTable
-import com.madrid.data.dataSource.remote.dto.genre.MovieGenre
+import com.madrid.domain.entity.Genre
 
-
-fun MovieGenre.toMovieGenreTable(): MovieGenreTable {
-    return MovieGenreTable(
-        genreId = this.id ?: 0,
-        genreTitle = this.name ?: "Unknown",
-        searchCount = 0
+fun MovieGenreTable.toGenre(): Genre {
+    return Genre(
+        id = this.genreId,
+        name = this.genreTitle,
+        interestPoints = this.searchCount
     )
 }
 
-fun MovieGenre.toSeriesGenreTable(): SeriesGenreTable {
-    return SeriesGenreTable(
-        genreId = this.id ?: 0,
-        genreTitle = this.name ?: "Unknown",
-        searchCount = 0
-    )
-}
+
