+++ domain/src/main/java/com/madrid/domain/entity/Genre.kt
@@ -1,0 +1,7 @@
+package com.madrid.domain.entity
+
+data class Genre(
+    val id: Int,
+    val name: String,
+    val interestPoints: Int = 0
+)
